requirements.txt
diffusers
transformers
torch
accelerate
safetensors
matplotlib
pillow
tqdm
ipywidgets
xformers
